more webpack...

+++ update_repos.sh
@@ -67,28 +67,30 @@ fi
 npm i
 cd ..
 
-echo "UPDATING API..."
-cd cables_api
-if [ -n "${1}" ] && [ "clean" != "${1}" ]; then
-	git checkout "${1}"
+if [ -d cables_api ]; then
+  echo "UPDATING API..."
+  cd cables_api
+  if [ -n "${1}" ] && [ "clean" != "${1}" ]; then
+    git checkout "${1}"
+  fi
+  # get current branch
+  branch=`git rev-parse --abbrev-ref HEAD`
+  # ignore errors here, since branch might not be on remote
+  git fetch || true
+  git pull origin "$branch" || true
+  # merge current remote develop if branch is not master
+  if [ "master" != "${branch}" ]; then
+      echo "merging current state of origin/develop into ${branch}";
+      git merge origin/develop;
+  else
+      echo "WARNING: not merging origin/develop into master!"
+  fi
+  if [ "clean" == "${1}" ]; then
+    rm -rf node_modules/
+  fi
+  npm i
+  cd ..
 fi
-# get current branch
-branch=`git rev-parse --abbrev-ref HEAD`
-# ignore errors here, since branch might not be on remote
-git fetch || true
-git pull origin "$branch" || true
-# merge current remote develop if branch is not master
-if [ "master" != "${branch}" ]; then
-    echo "merging current state of origin/develop into ${branch}";
-    git merge origin/develop;
-else
-    echo "WARNING: not merging origin/develop into master!"
-fi
-if [ "clean" == "${1}" ]; then
-	rm -rf node_modules/
-fi
-npm i
-cd ..
 
 echo "UPDATING UI..."
 cd cables_ui
@@ -114,26 +116,28 @@ npm i
 cd ..
 echo "DONE"
 
-echo "UPDATING ELECTRON..."
-cd cables_electron
-if [ -n "${1}" ] && [ "clean" != "${1}" ]; then
-	git checkout "${1}"
+if [ -d cables_electron ]; then
+  echo "UPDATING ELECTRON..."
+  cd cables_electron
+  if [ -n "${1}" ] && [ "clean" != "${1}" ]; then
+    git checkout "${1}"
+  fi
+  # get current branch
+  branch=`git rev-parse --abbrev-ref HEAD`
+  # ignore errors here, since branch might not be on remote
+  git fetch || true
+  git pull origin "${branch}" || true
+  # merge current remote develop if branch is not master
+  if [ "master" != "${branch}" ]; then
+      echo "merging current state of origin/develop into ${branch}";
+      git merge origin/develop;
+  else
+      echo "WARNING: not merging origin/develop into master!"
+  fi
+  if [ "clean" == "${1}" ]; then
+    rm -rf node_modules/
+  fi
+  npm i
+  cd ..
 fi
-# get current branch
-branch=`git rev-parse --abbrev-ref HEAD`
-# ignore errors here, since branch might not be on remote
-git fetch || true
-git pull origin "${branch}" || true
-# merge current remote develop if branch is not master
-if [ "master" != "${branch}" ]; then
-    echo "merging current state of origin/develop into ${branch}";
-    git merge origin/develop;
-else
-    echo "WARNING: not merging origin/develop into master!"
-fi
-if [ "clean" == "${1}" ]; then
-	rm -rf node_modules/
-fi
-npm i
-cd ..
 echo "DONE"
